requirements.txt
python-dateutil==2.6.0
requests==2.13.0
pymongo==3.4.0
numpy==1.12.1
gevent==1.2.1
h5py==2.7.0
tensorflow
bs4==0.0.1
dill==0.2.7.1
raven==6.1.0
sklearn
scipy==0.19.1
lightgbm==2.0.4
flask==0.12.2
pandas
pytest
xgboost
statsmodels
